AsciiDoctor Documentation
This starts the conversion of the Wiki documents to
AsciiDoctor files.  This is a proof of concept to show
some of the capabilites of AsciiDoctor and how it can
potentially make some chores easier to do with the
documentation.  Plus it potentially allows the docs to
be published in a variety of formats like HTML, PDF, ePub,
Man, latex, or other formats.