James Cameron Avatar 2010 Instruction Card
This updates the README.md to include instructions for setting
up the DMD to display the instruction card for the Avatar table
as the game is launched.   This goes to the DMD as currently there
is already a backglass.png file included in the repo.

Instruction card not included as is hosted on VPUniverse.