Big Trouble Little China POV and Resolution Update
This updates the POV to use the Window mode (ViewCabMode 2).  It
uses LoadWeapon's POV for cabinent mode.  Rotation applied to the
playfield to get the field oriented correctly.

In addition, updated BackBufferScale to one level higher, and
added AAFactor to help with fps at the higher resolution.

Overall the new POV applies a more realistic view and helps with
that crazy first plunger ball shot to make it a bit more realistic
speed wise.

Included link to LoadWeapon's POV in the ini file.